chore: refactor configuration page to use modular HTML components for improved maintainability
