+++ templates/configPage.ts
@@ -2,8 +2,20 @@
  * Shared templates for the configuration page
  *
  * This file contains HTML templates that can be used by Cloudflare and future implementations.
+ * It serves as an integration point for the smaller, more maintainable component files.
  */
 import { getMDBListSearchFormHTML } from './mdblistTemplates';
+import { getConfigPageHeaderHTML } from './configPageHeader';
+import { getAddCatalogFormHTML } from './addCatalogComponent';
+import { getMDBListApiConfigHTML } from './mdblistComponent';
+import { getCatalogListHTML } from './catalogListComponent';
+import { getAddonInstallationHTML } from './addonInstallComponent';
+import { getSponsorBannerHTML } from './sponsorComponent';
+import { getFooterHTML } from './footerComponent';
+import { getPageScriptsHTML } from './pageScriptsComponent';
+import { getHTMLHead, getBodyOpeningHTML, getBodyClosingHTML } from './pageStructureComponents';
+import { getHomePageHTML } from './homePageTemplate';
+export { convertStremioUrl } from './utilities';
 
 /**
  * Creates the HTML for the configuration page
@@ -18,812 +30,29 @@ export function getConfigPageHTML(
   packageVersion: string = '1.0.0',
   apiKey: string = ''
 ) {
-  const catalogRows = catalogs
-    .map((catalog, index, array) => {
-      const isFirst = index === 0;
-      const isLast = index === array.length - 1;
-      const isOnly = array.length === 1;
-
-      return `
-    <div class="catalog-item flex items-start gap-2 sm:gap-4" data-draggable="true" data-catalog-id="${catalog.id}" data-catalog-index="${index}">
-      <span class="catalog-handle inline-flex items-center justify-center rounded-full bg-primary/10 text-primary h-7 w-7 sm:h-8 sm:w-8 text-sm sm:text-base font-semibold shrink-0 mt-1 md:cursor-grab">
-        <span>${index + 1}</span>
-        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="hidden mr-1"><path d="M12 5v14M5 12h14"></path></svg>
-      </span>
-      <div class="flex flex-col space-y-2 p-3 sm:p-4 rounded-lg bg-card border border-border hover:bg-accent/50 transition-colors flex-grow min-w-0">
-        <div class="flex flex-col md:flex-row md:justify-between md:items-start gap-2 sm:gap-3">
-          <div class="flex-grow overflow-hidden min-w-0">
-            <h3 class="font-medium break-words d-inline text-base sm:text-lg truncate">${catalog.name}</h3>
-            <p class="text-xs sm:text-sm text-muted-foreground break-words truncate">${catalog.id}</p>
-          </div>
-          <div class="flex flex-wrap items-center gap-2 shrink-0">
-            ${
-              !isFirst && !isOnly
-                ? `
-            <form method="POST" action="/configure/${userId}/moveUp" class="flex-shrink-0 md:hidden">
-              <input type="hidden" name="catalogId" value="${catalog.id}">
-              <button type="submit" class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-secondary text-secondary-foreground hover:bg-secondary/80 h-9 w-9 p-2">
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-chevron-up"><path d="m18 15-6-6-6 6"/></svg>
-                <span class="sr-only">Move Up</span>
-              </button>
-            </form>
-            `
-                : ''
-            }
-            ${
-              !isLast && !isOnly
-                ? `
-            <form method="POST" action="/configure/${userId}/moveDown" class="flex-shrink-0 md:hidden">
-              <input type="hidden" name="catalogId" value="${catalog.id}">
-              <button type="submit" class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-secondary text-secondary-foreground hover:bg-secondary/80 h-9 w-9 p-2">
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="lucide lucide-chevron-down"><path d="m6 9 6 6 6-6"/></svg>
-                <span class="sr-only">Move Down</span>
-              </button>
-            </form>
-            `
-                : ''
-            }
-            <form method="POST" action="/configure/${userId}/remove" class="flex-shrink-0">
-              <input type="hidden" name="catalogId" value="${catalog.id}">
-              <button type="submit" class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-destructive text-destructive-foreground hover:bg-destructive/90 h-9 px-4 py-2">
-                Remove
-              </button>
-            </form>
-          </div>
-        </div>
-        <p class="text-xs sm:text-sm text-muted-foreground break-words truncate">${catalog.description}</p>
-        <p class="text-xs font-mono text-muted-foreground truncate">${catalog.endpoint}</p>
-      </div>
-    </div>
-  `;
-    })
-    .join('');
-
-  // Create URLs for the Stremio integration with query parameters
-  // The baseUrl must already contain the protocol
-  const paramsObject = { userId };
-  const encodedParams = encodeURIComponent(JSON.stringify(paramsObject));
-  const stremioUrl = `stremio://${baseUrl.replace(/^https?:\/\//, '')}/${encodedParams}/manifest.json`;
-  const manifestUrl = `${baseUrl}/${encodedParams}/manifest.json`;
-  const qrCodeUrl = `https://api.qrserver.com/v1/create-qr-code/?size=150x150&data=${encodeURIComponent(
-    stremioUrl
-  )}`;
-
-  // Add MDBList search form
+  // Add MDBList search form if API key is available
   const mdblistSearchForm = apiKey ? getMDBListSearchFormHTML(userId) : '';
 
-  // Create MDBList API config section
-  const mdblistApiConfigSection = `
-    <section>
-      <div class="rounded-lg border bg-card p-6 shadow-sm mt-8">
-        <h2 class="text-xl font-semibold mb-4">MDBList API Configuration</h2>
-        <form method="POST" action="/configure/${userId}/mdblist/config">
-          <div class="grid gap-4">
-            <div class="grid gap-2">
-              <label for="apiKey" class="text-sm font-medium">MDBList API Key</label>
-              <input
-                type="text"
-                id="apiKey"
-                name="apiKey"
-                value="${apiKey}"
-                placeholder="Enter your MDBList API key..."
-                class="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-              />
-              <p class="text-sm text-muted-foreground">
-                Get your free API key from <a href="https://mdblist.com/preferences/" target="_blank" class="text-primary hover:underline">MDBList Preferences</a>.
-              </p>
-            </div>
-            <button
-              type="submit"
-              class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-primary text-primary-foreground hover:bg-primary/90 h-10 px-4 py-2"
-            >
-              Save API Configuration
-            </button>
-          </div>
-        </form>
-      </div>
-    </section>
-  `;
-
+  // Build the page by combining the components
   return `
-    <!DOCTYPE html>
-    <html lang="en" class="dark">
-      <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-        <title>AIOCatalogs - Configuration</title>
-        <link rel="icon" href="https://i.imgur.com/fRPYeIV.png" />
-        <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/toastify-js/src/toastify.min.css">
-        <script src="https://cdn.tailwindcss.com"></script>
-        <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/toastify-js"></script>
-        <script>
-          tailwind.config = {
-            darkMode: 'class',
-            theme: {
-              container: {
-                center: true,
-                padding: '2rem',
-                screens: {
-                  '2xl': '1400px',
-                },
-              },
-              extend: {
-                colors: {
-                  border: 'hsl(240 3.7% 15.9%)',
-                  input: 'hsl(240 3.7% 15.9%)',
-                  ring: 'hsl(142.1 70.6% 45.3%)',
-                  background: 'hsl(240 10% 3.9%)',
-                  foreground: 'hsl(0 0% 98%)',
-                  primary: {
-                    DEFAULT: 'hsl(142.1 70.6% 45.3%)',
-                    foreground: 'hsl(144.9 80.4% 10%)',
-                  },
-                  secondary: {
-                    DEFAULT: 'hsl(240 3.7% 15.9%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                  destructive: {
-                    DEFAULT: 'hsl(0 62.8% 30.6%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                  muted: {
-                    DEFAULT: 'hsl(240 3.7% 15.9%)',
-                    foreground: 'hsl(240 5% 64.9%)',
-                  },
-                  accent: {
-                    DEFAULT: 'hsl(240 3.7% 15.9%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                  card: {
-                    DEFAULT: 'hsl(240 10% 5.9%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                },
-                borderRadius: {
-                  lg: '0.5rem',
-                  md: 'calc(0.5rem - 2px)',
-                  sm: 'calc(0.5rem - 4px)',
-                },
-              },
-            },
-          };
-        </script>
-        <style>
-          @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
-
-          * {
-            font-family: 'Inter', sans-serif;
-          }
-
-          .bg-card-pattern {
-            background-color: hsla(240, 10%, 5.9%, 1);
-            background-image:
-              radial-gradient(at 67% 27%, hsla(215, 98%, 61%, 0.05) 0px, transparent 50%),
-              radial-gradient(at 0% 0%, hsla(343, 100%, 76%, 0.05) 0px, transparent 50%);
-          }
-        </style>
-      </head>
-      <body class="min-h-screen bg-background text-foreground bg-card-pattern" data-user-id="${userId}">
-        <div class="container py-10">
-          <header class="mb-10">
-            <div class="flex flex-col">
-              <div class="flex items-center justify-between mb-2">
-                <h1 class="text-3xl font-bold tracking-tight">AIOCatalogs - Configuration</h1>
-                <div class="hidden md:flex items-center space-x-3">
-                  <a
-                    href="/configure?noRedirect=true"
-                    class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-secondary text-secondary-foreground hover:bg-secondary/80 h-9 px-4 py-2"
-                  >
-                  <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="mr-2">
-                    <path d="m15 18-6-6 6-6"></path>
-                  </svg>
-                    Back to User Selection
-                  </a>
-                </div>
-              </div>
-              <p class="text-lg text-muted-foreground">
-                User ID: <span class="font-medium text-primary">${userId}</span>
-                <button 
-                  id="clearStoredUserBtn" 
-                  class="ml-2 inline-flex items-center justify-center whitespace-nowrap rounded-md text-xs font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-destructive text-destructive-foreground hover:bg-destructive/90 h-6 px-2 py-1"
-                >
-                  Clear Config from Local Storage
-                </button>
-              </p>
-            </div>
-          </header>
-
-          <!-- Back to User Selection button, mobile friendly -->
-          <div class="md:hidden mb-6">
-            <a
-              href="/configure?noRedirect=true"
-              class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-secondary text-secondary-foreground hover:bg-secondary/80 h-10 px-4 py-2 w-full"
-            >
-              <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="mr-2">
-                <path d="m15 18-6-6 6-6"></path>
-              </svg>
-              Back to User Selection
-            </a>
-          </div>
-
-          <div id="notifications"></div>
-
-          <!-- Sponsor-Button -->
-          <div class="mb-8 p-4 rounded-lg bg-yellow-500/10 border border-yellow-500 flex flex-col sm:flex-row items-center justify-between">
-            <div>
-              <h3 class="font-semibold mb-1">Support AIOCatalogs</h3>
-              <p class="text-sm text-muted-foreground">Help keep this project maintained and running</p>
-            </div>
-            <a 
-              href="https://buymeacoffee.com/pantel" 
-              target="_blank" 
-              rel="noopener noreferrer" 
-              class="mt-3 sm:mt-0 inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-yellow-500 text-black hover:bg-yellow-500/90 h-10 px-4 py-2"
-            >
-              <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="mr-2">
-                <path d="M5 12l-.854.854A4 4 0 0 0 11 17.236h2a4 4 0 0 0 2.761-1.104l.483-.483"></path>
-                <path d="M18 12v.01"></path>
-                <path d="M6 13v.01"></path>
-                <path d="M5 8v.01"></path>
-                <path d="M19 8v.01"></path>
-                <path d="M9 17c-2.5-1-4-3.5-4-7a8 8 0 0 1 8-8h.5a8 8 0 0 1 7.5 8.5c0 11-8 8-8 8"></path>
-              </svg>
-              Donate
-            </a>
-          </div>
-
-          <div class="grid gap-8">
-            <section>
-              <div class="rounded-lg border bg-card p-6 shadow-sm">
-                <h2 class="text-xl font-semibold mb-4">Add New Catalog</h2>
-                <form method="POST" action="/configure/${userId}/add">
-                  <div class="grid gap-4">
-                    <div class="grid gap-2">
-                      <label for="catalogUrl" class="text-sm font-medium"
-                        >Catalog Manifest URL</label
-                      >
-                      <input
-                        type="url"
-                        id="catalogUrl"
-                        name="catalogUrl"
-                        placeholder="https://example.com/manifest.json"
-                        required
-                        class="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-                      />
-                      <p class="text-sm text-muted-foreground">
-                        Enter the manifest.json URL of the catalog addon you want to add (e.g.
-                        MDBList)
-                      </p>
-                    </div>
-                    <button
-                      type="submit"
-                      class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-primary text-primary-foreground hover:bg-primary/90 h-10 px-4 py-2 w-full sm:w-auto"
-                    >
-                      Add Catalog
-                    </button>
-                  </div>
-                </form>
-              </div>
-            </section>
-
-            ${mdblistApiConfigSection}
-
-            ${mdblistSearchForm}
-
-            <section>
-              <h2 class="text-xl font-semibold mb-4">Your Catalogs</h2>
-              ${
-                catalogs.length === 0
-                  ? '<div class="p-4 rounded-lg bg-secondary/50 text-muted-foreground">No catalogs added yet. Add some catalogs to get started.</div>'
-                  : `<div id="catalog-list" class="grid grid-cols-1 gap-4">${catalogRows}</div>`
-              }
-              ${
-                catalogs.length > 1
-                  ? `<div class="mt-4 p-4 rounded-lg bg-secondary/50 border-l-4 border-primary">
-                      <p class="text-sm text-muted-foreground">
-                        <strong>Hint:</strong> <span class="md:inline hidden">You can reorder the catalogs by dragging and dropping them.</span>
-                        <span class="md:hidden">Use the arrow buttons to reorder catalogs.</span>
-                      </p>
-                    </div>`
-                  : ''
-              }
-            </section>
-
-            <section class="mt-6">
-              <div class="rounded-lg border bg-card p-6 shadow-sm">
-                <h2 class="text-xl font-semibold mb-4">Install Your Addon</h2>
-
-                <div class="grid grid-cols-1 md:grid-cols-2 gap-8">
-                  <div>
-                    <p class="mb-4">
-                      Click the button below to install this addon with your catalogs in Stremio:
-                    </p>
-                    <a
-                      href="${stremioUrl}"
-                      class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-primary text-primary-foreground hover:bg-primary/90 h-10 px-4 py-2"
-                    >
-                      Add to Stremio
-                    </a>
-
-                    <div class="mt-6">
-                      <p class="mb-2">Alternatively, add this URL manually:</p>
-                      <div class="p-3 rounded-md bg-muted font-mono text-sm break-all">
-                        ${manifestUrl}
-                      </div>
-                      <button
-                        class="mt-2 inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 border border-input bg-background hover:bg-accent hover:text-accent-foreground h-9 px-4 py-2"
-                        type="button"
-                        onclick="navigator.clipboard.writeText('${manifestUrl}')"
-                      >
-                        Copy URL
-                      </button>
-                    </div>
-                  </div>
-
-                  <div class="flex flex-col items-center">
-                    <p class="mb-4">Or scan this QR code with your mobile device:</p>
-                    <div class="p-4 bg-white rounded-lg inline-block">
-                      <img
-                        src="${qrCodeUrl}"
-                        alt="QR Code for Stremio Addon"
-                        class="h-[150px] w-[150px]"
-                      />
-                    </div>
-                    <p class="text-sm text-muted-foreground mt-4 text-center">
-                      The QR code contains the Stremio URL for this addon. Scan it with your
-                      smartphone to install the addon in the mobile Stremio app.
-                    </p>
-                  </div>
-                </div>
-
-                <div class="mt-4 p-4 rounded-lg bg-secondary/50 border-l-4 border-primary">
-                  <p class="text-sm text-muted-foreground">
-                    <strong>Note:</strong> You have to reinstall the addon after adding or removing
-                    catalogs.
-                  </p>
-                </div>
-              </div>
-            </section>
-          </div>
-        </div>
-
-        <!-- Footer with Social Links -->
-        <footer class="py-6 border-t border-border mt-10">
-          <div class="container flex flex-col items-center gap-4">
-            <div class="flex space-x-4">
-              <!-- GitHub -->
-              <a href="https://github.com/panteLx/aiocatalogs" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <path d="M15 22v-4a4.8 4.8 0 0 0-1-3.5c3 0 6-2 6-5.5.08-1.25-.27-2.48-1-3.5.28-1.15.28-2.35 0-3.5 0 0-1 0-3 1.5-2.64-.5-5.36-.5-8 0C6 2 5 2 5 2c-.3 1.15-.3 2.35 0 3.5A5.403 5.403 0 0 0 4 9c0 3.5 3 5.5 6 5.5-.39.49-.68 1.05-.85 1.65-.17.6-.22 1.23-.15 1.85v4"></path>
-                  <path d="M9 18c-4.51 2-5-2-7-2"></path>
-                </svg>
-              </a>
-              
-              <!-- Discord -->
-              <a href="https://discord.gg/Ma4SnagqwE" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor"  stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z"></path>
-                </svg>
-              </a>
-              
-              <!-- Sponsor-Link -->
-              <a href="https://buymeacoffee.com/pantel" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <path d="M18 8h1a4 4 0 0 1 0 8h-1"></path>
-                  <path d="M2 8h16v9a4 4 0 0 1-4 4H6a4 4 0 0 1-4-4V8z"></path>
-                  <path d="M6 1v3"></path>
-                  <path d="M10 1v3"></path>
-                  <path d="M14 1v3"></path>
-                </svg>
-              </a>
-            </div>
-            <div class="text-sm text-muted-foreground text-center">
-              <p>Version: v${packageVersion} - Developed by panteLx</p>
-              <p class="mt-1">&copy; ${new Date().getFullYear()} AIOCatalogs - <a href="https://github.com/panteLx/aiocatalogs/blob/main/LICENSE" target="_blank" rel="noopener noreferrer" class="underline hover:text-primary">MIT License</a></p>
-            </div>
-          </div>
-        </footer>
-
-        <script>
-          function copyToClipboard(text) {
-            navigator.clipboard.writeText(text);
-            showToast('Copied to clipboard!', 'success');
-          }
-
-          // Toast notification function
-          function showToast(message, type) {
-            const bgColor = type === 'success' ? 'linear-gradient(to right, #0cce6b, #0caa57)' : 
-                           type === 'error' ? 'linear-gradient(to right, #e53935, #c62828)' : 
-                           'linear-gradient(to right, #2196f3, #1976d2)';
-            
-            Toastify({
-              text: message,
-              duration: 3000,
-              gravity: "top",
-              position: "right",
-              style: {
-                background: bgColor,
-                borderRadius: "8px",
-              },
-              stopOnFocus: true,
-            }).showToast();
-          }
-
-          // Save user ID to localStorage when page loads
-          document.addEventListener('DOMContentLoaded', function() {
-            // Get user ID from data attribute
-            const userId = document.body.getAttribute('data-user-id');
-            console.log("Storing userId in localStorage:", userId);
-            localStorage.setItem('aioCatalogsUserId', userId);
-            
-            // Set up clear stored user button
-            document.getElementById('clearStoredUserBtn').addEventListener('click', function() {
-              localStorage.removeItem('aioCatalogsUserId');
-              showToast('Stored user ID cleared. Next time you visit, you\\'ll need to enter it manually.', 'success');
-            });
-            
-            // Show toasts if there are messages or errors
-            ${message ? `showToast("${message}", "success");` : ''}
-            ${error ? `showToast("${error}", "error");` : ''}
-
-            // Check if it's a mobile device
-            const isMobile = window.matchMedia("(max-width: 768px)").matches;
-            
-            // Set up drag and drop functionality only for desktop devices
-            if (!isMobile) {
-              setupDragAndDrop(userId);
-              
-              // Make items draggable on desktop
-              document.querySelectorAll('[data-draggable="true"]').forEach(item => {
-                item.setAttribute('draggable', 'true');
-              });
-            }
-          });
-          
-          // Function to set up drag and drop
-          function setupDragAndDrop(userId) {
-            const catalogList = document.getElementById('catalog-list');
-            if (catalogList) {
-              let draggedItem = null;
-              let sourceIndex = -1;
-              
-              // Add event listeners to all catalog items
-              const catalogItems = document.querySelectorAll('.catalog-item');
-              catalogItems.forEach((item) => {
-                // Drag start
-                item.addEventListener('dragstart', function(e) {
-                  draggedItem = this;
-                  sourceIndex = parseInt(this.getAttribute('data-catalog-index'));
-                  // Set a timeout to add opacity class
-                  setTimeout(() => {
-                    this.classList.add('opacity-50');
-                  }, 0);
-                });
-                
-                // Drag end
-                item.addEventListener('dragend', function() {
-                  this.classList.remove('opacity-50');
-                  draggedItem = null;
-                  sourceIndex = -1;
-                  // Update the index numbers visually
-                  updateIndexNumbers();
-                });
-                
-                // Drag over
-                item.addEventListener('dragover', function(e) {
-                  e.preventDefault();
-                  if (draggedItem === this) return;
-                  this.classList.add('bg-accent');
-                });
-                
-                // Drag leave
-                item.addEventListener('dragleave', function() {
-                  this.classList.remove('bg-accent');
-                });
-                
-                // Drag drop
-                item.addEventListener('drop', async function(e) {
-                  e.preventDefault();
-                  this.classList.remove('bg-accent');
-                  
-                  if (draggedItem === this) return;
-                  
-                  const targetIndex = parseInt(this.getAttribute('data-catalog-index'));
-                  const draggedId = draggedItem.getAttribute('data-catalog-id');
-                  
-                  // If dragging downwards, insert after the target item
-                  if (sourceIndex < targetIndex) {
-                    catalogList.insertBefore(draggedItem, this.nextSibling);
-                  } 
-                  // If dragging upwards, insert before the target item
-                  else {
-                    catalogList.insertBefore(draggedItem, this);
-                  }
-                  
-                  // Update the backend using the existing moveUp/moveDown functions
-                  await updateCatalogOrder(draggedId, sourceIndex, targetIndex, userId);
-                });
-              });
-            }
-          }
-          
-          // Function to update backend catalog order
-          async function updateCatalogOrder(catalogId, fromIndex, toIndex, userId) {
-            const steps = Math.abs(toIndex - fromIndex);
-            const moveDirection = fromIndex < toIndex ? 'down' : 'up';
-            const moveEndpoint = moveDirection === 'down' ? 'moveDown' : 'moveUp';
-            
-            try {
-              // Make the necessary number of move requests to get from source to target
-              for (let i = 0; i < steps; i++) {
-                const formData = new FormData();
-                formData.append('catalogId', catalogId);
-                
-                await fetch('/configure/' + userId + '/' + moveEndpoint, {
-                  method: 'POST',
-                  body: formData
-                });
-              }
-              
-              showToast('Catalog position updated successfully', 'success');
-              
-              // Update the data-catalog-index attributes
-              updateCatalogIndexAttributes();
-              
-            } catch (error) {
-              console.error('Error updating catalog position:', error);
-              showToast('Failed to update catalog position', 'error');
-              
-              // If error, reload the page to restore the correct order
-              window.location.reload();
-            }
-          }
-          
-          // Function to update the numbered indicators
-          function updateIndexNumbers() {
-            document.querySelectorAll('.catalog-item').forEach((item, index) => {
-              const indicator = item.querySelector('.catalog-handle span');
-              if (indicator) {
-                indicator.textContent = (index + 1).toString();
-              }
-            });
-          }
-          
-          // Function to update data-catalog-index attributes
-          function updateCatalogIndexAttributes() {
-            document.querySelectorAll('.catalog-item').forEach((item, index) => {
-              item.setAttribute('data-catalog-index', index.toString());
-            });
-          }
-        </script>
-      </body>
-    </html>
+    ${getHTMLHead('AIOCatalogs - Configuration')}
+    ${getBodyOpeningHTML(userId)}
+    ${getConfigPageHeaderHTML(userId)}
+    ${getSponsorBannerHTML()}
+    
+    <div class="grid gap-8">
+      ${getAddCatalogFormHTML(userId)}
+      ${apiKey ? getMDBListApiConfigHTML(userId, apiKey) : ''}
+      ${mdblistSearchForm}
+      ${getCatalogListHTML(userId, catalogs)}
+      ${getAddonInstallationHTML(userId, baseUrl)}
+    </div>
+    
+    ${getFooterHTML(packageVersion)}
+    ${getPageScriptsHTML(message, error)}
+    ${getBodyClosingHTML()}
   `;
 }
 
-/**
- * Creates the HTML for the home page
- */
-export function getHomePageHTML(
-  message: string = '',
-  error: string = '',
-  packageVersion: string = '1.0.0'
-) {
-  return `
-    <!DOCTYPE html>
-    <html lang="en" class="dark">
-      <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>AIOCatalogs</title>
-        <link rel="icon" href="https://i.imgur.com/fRPYeIV.png" />
-        <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/toastify-js/src/toastify.min.css">
-        <script src="https://cdn.tailwindcss.com"></script>
-        <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/toastify-js"></script>
-        <script>
-          tailwind.config = {
-            darkMode: 'class',
-            theme: {
-              container: {
-                center: true,
-                padding: '2rem',
-                screens: {
-                  '2xl': '1400px',
-                },
-              },
-              extend: {
-                colors: {
-                  border: 'hsl(240 3.7% 15.9%)',
-                  input: 'hsl(240 3.7% 15.9%)',
-                  ring: 'hsl(142.1 70.6% 45.3%)',
-                  background: 'hsl(240 10% 3.9%)',
-                  foreground: 'hsl(0 0% 98%)',
-                  primary: {
-                    DEFAULT: 'hsl(142.1 70.6% 45.3%)',
-                    foreground: 'hsl(144.9 80.4% 10%)',
-                  },
-                  secondary: {
-                    DEFAULT: 'hsl(240 3.7% 15.9%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                  destructive: {
-                    DEFAULT: 'hsl(0 62.8% 30.6%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                  muted: {
-                    DEFAULT: 'hsl(240 3.7% 15.9%)',
-                    foreground: 'hsl(240 5% 64.9%)',
-                  },
-                  accent: {
-                    DEFAULT: 'hsl(240 3.7% 15.9%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                  card: {
-                    DEFAULT: 'hsl(240 10% 5.9%)',
-                    foreground: 'hsl(0 0% 98%)',
-                  },
-                },
-                borderRadius: {
-                  lg: '0.5rem',
-                  md: 'calc(0.5rem - 2px)',
-                  sm: 'calc(0.5rem - 4px)',
-                },
-              },
-            },
-          };
-        </script>
-        <style>
-          @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
-
-          * {
-            font-family: 'Inter', sans-serif;
-          }
-
-          .bg-card-pattern {
-            background-color: hsla(240, 10%, 5.9%, 1);
-            background-image:
-              radial-gradient(at 67% 27%, hsla(215, 98%, 61%, 0.05) 0px, transparent 50%),
-              radial-gradient(at 0% 0%, hsla(343, 100%, 76%, 0.05) 0px, transparent 50%);
-          }
-        </style>
-      </head>
-      <body class="min-h-screen bg-background text-foreground bg-card-pattern">
-        <div class="container py-16">
-          <h1 class="text-4xl font-bold tracking-tight text-center mb-4">
-            All-in-One Catalogs - User Selection
-          </h1>
-
-          <div id="notifications"></div>
-
-          <div class="grid gap-8 md:grid-cols-2 max-w-3xl mx-auto mt-20">
-            <div class="rounded-lg border bg-card p-6 shadow-sm">
-              <h2 class="text-xl font-semibold mb-4">New User</h2>
-              <p class="text-muted-foreground mb-6">
-                Create a new configuration for your catalogs.
-              </p>
-              <form method="POST" action="/configure/create">
-                <button
-                  type="submit"
-                  class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-primary text-primary-foreground hover:bg-primary/90 h-10 px-4 py-2 w-full"
-                >
-                  Create New Configuration
-                </button>
-              </form>
-            </div>
-
-            <div class="rounded-lg border bg-card p-6 shadow-sm">
-              <h2 class="text-xl font-semibold mb-4">Existing User</h2>
-              <p class="text-muted-foreground mb-4">Load your existing catalog configuration.</p>
-              <form method="POST" action="/configure/load">
-                <div class="grid gap-4">
-                  <div class="grid gap-2">
-                    <label for="userId" class="text-sm font-medium">Your User ID</label>
-                    <input
-                      type="text"
-                      id="userId"
-                      name="userId"
-                      required
-                      class="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-                    />
-                  </div>
-                  <button
-                    type="submit"
-                    class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-secondary text-secondary-foreground hover:bg-secondary/80 h-10 px-4 py-2 w-full"
-                  >
-                    Load Configuration
-                  </button>
-                </div>
-              </form>
-            </div>
-          </div>
-        </div>
-
-        <!-- Footer with Social Links -->
-        <footer class="py-6 border-t border-border mt-10">
-          <div class="container flex flex-col items-center gap-4">
-            <div class="flex space-x-4">
-              <!-- GitHub -->
-              <a href="https://github.com/panteLx/aiocatalogs" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <path d="M15 22v-4a4.8 4.8 0 0 0-1-3.5c3 0 6-2 6-5.5.08-1.25-.27-2.48-1-3.5.28-1.15.28-2.35 0-3.5 0 0-1 0-3 1.5-2.64-.5-5.36-.5-8 0C6 2 5 2 5 2c-.3 1.15-.3 2.35 0 3.5A5.403 5.403 0 0 0 4 9c0 3.5 3 5.5 6 5.5-.39.49-.68 1.05-.85 1.65-.17.6-.22 1.23-.15 1.85v4"></path>
-                  <path d="M9 18c-4.51 2-5-2-7-2"></path>
-                </svg>
-              </a>
-              
-              <!-- Discord -->
-              <a href="https://discord.gg/Ma4SnagqwE" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor"  stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z"></path>
-                </svg>
-              </a>
-              
-              <!-- Sponsor-Link -->
-              <a href="https://buymeacoffee.com/pantel" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
-                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                  <path d="M18 8h1a4 4 0 0 1 0 8h-1"></path>
-                  <path d="M2 8h16v9a4 4 0 0 1-4 4H6a4 4 0 0 1-4-4V8z"></path>
-                  <path d="M6 1v3"></path>
-                  <path d="M10 1v3"></path>
-                  <path d="M14 1v3"></path>
-                </svg>
-              </a>
-            </div>
-            <div class="text-sm text-muted-foreground text-center">
-              <p>Version: v${packageVersion} - Developed by panteLx</p>
-              <p class="mt-1">&copy; ${new Date().getFullYear()} AIOCatalogs - <a href="https://github.com/panteLx/aiocatalogs/blob/main/LICENSE" target="_blank" rel="noopener noreferrer" class="underline hover:text-primary">MIT License</a></p>
-            </div>
-          </div>
-        </footer>
-
-        <script>
-          // Toast notification function
-          function showToast(message, type) {
-            const bgColor = type === 'success' ? 'linear-gradient(to right, #0cce6b, #0caa57)' : 
-                           type === 'error' ? 'linear-gradient(to right, #e53935, #c62828)' : 
-                           'linear-gradient(to right, #2196f3, #1976d2)';
-            
-            Toastify({
-              text: message,
-              duration: 3000,
-              gravity: "top",
-              position: "right",
-              style: {
-                background: bgColor,
-                borderRadius: "8px",
-              },
-              stopOnFocus: true,
-            }).showToast();
-          }
-
-          // Show toasts if there are messages or errors and set up saved ID
-          document.addEventListener('DOMContentLoaded', function() {
-            ${message ? `showToast("${message}", "success");` : ''}
-            ${error ? `showToast("${error}", "error");` : ''}
-            
-            // Set up the user ID field with the saved value if available
-            console.log("Checking for saved user ID in localStorage");
-            const savedUserId = localStorage.getItem('aioCatalogsUserId');
-            console.log("Saved user ID from localStorage:", savedUserId);
-            
-            const userIdField = document.getElementById('userId');
-            
-            if (savedUserId && savedUserId.trim().length > 0) {
-              console.log("Setting userId field value:", savedUserId);
-              userIdField.value = savedUserId;
-            }
-          });
-        </script>
-      </body>
-    </html>
-  `;
-}
-
-/**
- * Helper function to convert stremio:// URLs to https://
- */
-export function convertStremioUrl(url: string): string {
-  if (url.startsWith('stremio://')) {
-    return url.replace('stremio://', 'https://');
-  }
-  return url;
-}
+// Re-export the home page HTML function
+export { getHomePageHTML };
