+++ templates/homePageTemplate.ts
@@ -1,0 +1,220 @@
+/**
+ * Home page template for the configuration page
+ */
+
+/**
+ * Creates the HTML for the home page
+ */
+export function getHomePageHTML(
+  message: string = '',
+  error: string = '',
+  packageVersion: string = '1.0.0'
+) {
+  return `
+    <!DOCTYPE html>
+    <html lang="en" class="dark">
+      <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>AIOCatalogs</title>
+        <link rel="icon" href="https://i.imgur.com/fRPYeIV.png" />
+        <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/toastify-js/src/toastify.min.css">
+        <script src="https://cdn.tailwindcss.com"></script>
+        <script type="text/javascript" src="https://cdn.jsdelivr.net/npm/toastify-js"></script>
+        <script>
+          tailwind.config = {
+            darkMode: 'class',
+            theme: {
+              container: {
+                center: true,
+                padding: '2rem',
+                screens: {
+                  '2xl': '1400px',
+                },
+              },
+              extend: {
+                colors: {
+                  border: 'hsl(240 3.7% 15.9%)',
+                  input: 'hsl(240 3.7% 15.9%)',
+                  ring: 'hsl(142.1 70.6% 45.3%)',
+                  background: 'hsl(240 10% 3.9%)',
+                  foreground: 'hsl(0 0% 98%)',
+                  primary: {
+                    DEFAULT: 'hsl(142.1 70.6% 45.3%)',
+                    foreground: 'hsl(144.9 80.4% 10%)',
+                  },
+                  secondary: {
+                    DEFAULT: 'hsl(240 3.7% 15.9%)',
+                    foreground: 'hsl(0 0% 98%)',
+                  },
+                  destructive: {
+                    DEFAULT: 'hsl(0 62.8% 30.6%)',
+                    foreground: 'hsl(0 0% 98%)',
+                  },
+                  muted: {
+                    DEFAULT: 'hsl(240 3.7% 15.9%)',
+                    foreground: 'hsl(240 5% 64.9%)',
+                  },
+                  accent: {
+                    DEFAULT: 'hsl(240 3.7% 15.9%)',
+                    foreground: 'hsl(0 0% 98%)',
+                  },
+                  card: {
+                    DEFAULT: 'hsl(240 10% 5.9%)',
+                    foreground: 'hsl(0 0% 98%)',
+                  },
+                },
+                borderRadius: {
+                  lg: '0.5rem',
+                  md: 'calc(0.5rem - 2px)',
+                  sm: 'calc(0.5rem - 4px)',
+                },
+              },
+            },
+          };
+        </script>
+        <style>
+          @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
+
+          * {
+            font-family: 'Inter', sans-serif;
+          }
+
+          .bg-card-pattern {
+            background-color: hsla(240, 10%, 5.9%, 1);
+            background-image:
+              radial-gradient(at 67% 27%, hsla(215, 98%, 61%, 0.05) 0px, transparent 50%),
+              radial-gradient(at 0% 0%, hsla(343, 100%, 76%, 0.05) 0px, transparent 50%);
+          }
+        </style>
+      </head>
+      <body class="min-h-screen bg-background text-foreground bg-card-pattern">
+        <div class="container py-16">
+          <h1 class="text-4xl font-bold tracking-tight text-center mb-4">
+            All-in-One Catalogs - User Selection
+          </h1>
+
+          <div id="notifications"></div>
+
+          <div class="grid gap-8 md:grid-cols-2 max-w-3xl mx-auto mt-20">
+            <div class="rounded-lg border bg-card p-6 shadow-sm">
+              <h2 class="text-xl font-semibold mb-4">New User</h2>
+              <p class="text-muted-foreground mb-6">
+                Create a new configuration for your catalogs.
+              </p>
+              <form method="POST" action="/configure/create">
+                <button
+                  type="submit"
+                  class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-primary text-primary-foreground hover:bg-primary/90 h-10 px-4 py-2 w-full"
+                >
+                  Create New Configuration
+                </button>
+              </form>
+            </div>
+
+            <div class="rounded-lg border bg-card p-6 shadow-sm">
+              <h2 class="text-xl font-semibold mb-4">Existing User</h2>
+              <p class="text-muted-foreground mb-4">Load your existing catalog configuration.</p>
+              <form method="POST" action="/configure/load">
+                <div class="grid gap-4">
+                  <div class="grid gap-2">
+                    <label for="userId" class="text-sm font-medium">Your User ID</label>
+                    <input
+                      type="text"
+                      id="userId"
+                      name="userId"
+                      required
+                      class="flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
+                    />
+                  </div>
+                  <button
+                    type="submit"
+                    class="inline-flex items-center justify-center whitespace-nowrap rounded-md text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 bg-secondary text-secondary-foreground hover:bg-secondary/80 h-10 px-4 py-2 w-full"
+                  >
+                    Load Configuration
+                  </button>
+                </div>
+              </form>
+            </div>
+          </div>
+        </div>
+
+        <!-- Footer with Social Links -->
+        <footer class="py-6 border-t border-border mt-10">
+          <div class="container flex flex-col items-center gap-4">
+            <div class="flex space-x-4">
+              <!-- GitHub -->
+              <a href="https://github.com/panteLx/aiocatalogs" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
+                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+                  <path d="M15 22v-4a4.8 4.8 0 0 0-1-3.5c3 0 6-2 6-5.5.08-1.25-.27-2.48-1-3.5.28-1.15.28-2.35 0-3.5 0 0-1 0-3 1.5-2.64-.5-5.36-.5-8 0C6 2 5 2 5 2c-.3 1.15-.3 2.35 0 3.5A5.403 5.403 0 0 0 4 9c0 3.5 3 5.5 6 5.5-.39.49-.68 1.05-.85 1.65-.17.6-.22 1.23-.15 1.85v4"></path>
+                  <path d="M9 18c-4.51 2-5-2-7-2"></path>
+                </svg>
+              </a>
+              
+              <!-- Discord -->
+              <a href="https://discord.gg/Ma4SnagqwE" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
+                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor"  stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+                  <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z"></path>
+                </svg>
+              </a>
+              
+              <!-- Sponsor-Link -->
+              <a href="https://buymeacoffee.com/pantel" target="_blank" rel="noopener noreferrer" class="text-muted-foreground hover:text-primary transition-colors">
+                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+                  <path d="M18 8h1a4 4 0 0 1 0 8h-1"></path>
+                  <path d="M2 8h16v9a4 4 0 0 1-4 4H6a4 4 0 0 1-4-4V8z"></path>
+                  <path d="M6 1v3"></path>
+                  <path d="M10 1v3"></path>
+                  <path d="M14 1v3"></path>
+                </svg>
+              </a>
+            </div>
+            <div class="text-sm text-muted-foreground text-center">
+              <p>Version: v${packageVersion} - Developed by panteLx</p>
+              <p class="mt-1">&copy; ${new Date().getFullYear()} AIOCatalogs - <a href="https://github.com/panteLx/aiocatalogs/blob/main/LICENSE" target="_blank" rel="noopener noreferrer" class="underline hover:text-primary">MIT License</a></p>
+            </div>
+          </div>
+        </footer>
+
+        <script>
+          // Toast notification function
+          function showToast(message, type) {
+            const bgColor = type === 'success' ? 'linear-gradient(to right, #0cce6b, #0caa57)' : 
+                           type === 'error' ? 'linear-gradient(to right, #e53935, #c62828)' : 
+                           'linear-gradient(to right, #2196f3, #1976d2)';
+            
+            Toastify({
+              text: message,
+              duration: 3000,
+              gravity: "top",
+              position: "right",
+              style: {
+                background: bgColor,
+                borderRadius: "8px",
+              },
+              stopOnFocus: true,
+            }).showToast();
+          }
+
+          // Show toasts if there are messages or errors and set up saved ID
+          document.addEventListener('DOMContentLoaded', function() {
+            ${message ? `showToast("${message}", "success");` : ''}
+            ${error ? `showToast("${error}", "error");` : ''}
+            
+            // Set up the user ID field with the saved value if available
+            console.log("Checking for saved user ID in localStorage");
+            const savedUserId = localStorage.getItem('aioCatalogsUserId');
+            console.log("Saved user ID from localStorage:", savedUserId);
+            
+            const userIdField = document.getElementById('userId');
+            
+            if (savedUserId && savedUserId.trim().length > 0) {
+              console.log("Setting userId field value:", savedUserId);
+              userIdField.value = savedUserId;
+            }
+          });
+        </script>
+      </body>
+    </html>
+  `;
+}
